fix: normalize delivery coordinates precision

+++ src/main/java/sv/edu/uca/delivery/backend/dto/DeliveryProfileResponse.java
@@ -1,14 +1,15 @@
 package sv.edu.uca.delivery.backend.dto;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record DeliveryProfileResponse(
         UUID deliveryUserId,
         String deliveryUserName,
         boolean available,
-        Double latitude,
-        Double longitude,
+        BigDecimal latitude,
+        BigDecimal longitude,
         LocalDateTime locationRecordedAt,
         Double averageRating,
         long reviewCount
